source renamed and revamped to source-shipstation
