+++ .vscode/settings.json
@@ -41,5 +41,6 @@
     "editor.codeActionsOnSave": {
       "source.fixAll.stylelint": true
     }
-  }
+  },
+  "python.analysis.extraPaths": ["./airbyte-cdk/python"]
 }
