+++ airbyte-integrations/connectors/source-shipstation/Dockerfile
@@ -29,13 +29,13 @@ RUN apk --no-cache add bash
 
 # copy payload code only
 COPY main.py ./
-COPY source_shipstation_abstract ./source_shipstation_abstract
+COPY source_shipstation ./source_shipstation
 
 ENV AIRBYTE_ENTRYPOINT "python /airbyte/integration_code/main.py"
 ENTRYPOINT ["python", "/airbyte/integration_code/main.py"]
 
 LABEL io.airbyte.version=0.1.0
-LABEL io.airbyte.name=airbyte/source-shipstation-abstract
+LABEL io.airbyte.name=airbyte/source-shipstation
 
 # Add Docker repository and image tag placeholders
 ARG DOCKER_REPOSITORY
