+++ airbyte-integrations/connectors/source-shipstation/README.md
@@ -6,23 +6,28 @@ For information about how to use this connector within Airbyte, see [the documen
 ## Local development
 
 ### Prerequisites
+
 **To iterate on this connector, make sure to complete this prerequisites section.**
 
 #### Minimum Python version required `= 3.9.0`
 
 #### Build & Activate Virtual Environment and install dependencies
+
 From this connector directory, create a virtual environment:
+
 ```
 python -m venv .venv
 ```
 
 This will generate a virtualenv for this module in `.venv/`. Make sure this venv is active in your
 development environment of choice. To activate it from the terminal, run:
+
 ```
 source .venv/bin/activate
 pip install -r requirements.txt
 pip install '.[tests]'
 ```
+
 If you are in an IDE, follow your IDE's instructions to activate the virtualenv.
 
 Note that while we are installing dependencies from `requirements.txt`, you should only edit `setup.py` for your dependencies. `requirements.txt` is
@@ -31,16 +36,19 @@ If this is mumbo jumbo to you, don't worry about it, just put your deps in `setu
 should work as you expect.
 
 #### Building via Gradle
+
 You can also build the connector in Gradle. This is typically used in CI and not needed for your development workflow.
 
 To build using Gradle, from the Airbyte repository root, run:
+
 ```
-./gradlew :airbyte-integrations:connectors:source-shipstation-abstract:build
+./gradlew :airbyte-integrations:connectors:source-shipstation:build
 ```
 
 #### Create credentials
-**If you are a community contributor**, follow the instructions in the [documentation](https://docs.airbyte.io/integrations/sources/shipstation-abstract)
-to generate the necessary credentials. Then create a file `secrets/config.json` conforming to the `source_shipstation_abstract/spec.yaml` file.
+
+**If you are a community contributor**, follow the instructions in the [documentation](https://docs.airbyte.io/integrations/sources/shipstation)
+to generate the necessary credentials. Then create a file `secrets/config.json` conforming to the `source_shipstation/spec.yaml` file.
 Note that any directory named `secrets` is gitignored across the entire Airbyte repo, so there is no danger of accidentally checking in sensitive information.
 See `integration_tests/sample_config.json` for a sample config file.
 
@@ -48,8 +56,8 @@ See `integration_tests/sample_config.json` for a sample config file.
 and place them into `secrets/config.json`.
 
 ### Locally running the connector
+
 ```
-python main.py spec
 python main.py check --config secrets/config.json
 python main.py discover --config secrets/config.json
 python main.py read --config secrets/config.json --catalog integration_tests/configured_catalog.json
@@ -58,73 +66,101 @@ python main.py read --config secrets/config.json --catalog integration_tests/con
 ### Locally running the connector docker image
 
 #### Build
+
 First, make sure you build the latest Docker image:
+
 ```
-docker build . -t airbyte/source-shipstation-abstract:dev
+docker build . -t airbyte/source-shipstation:dev
 ```
 
 You can also build the connector image via Gradle:
+
 ```
-./gradlew :airbyte-integrations:connectors:source-shipstation-abstract:airbyteDocker
+./gradlew :airbyte-integrations:connectors:source-shipstation:airbyteDocker
 ```
+
 When building via Gradle, the docker image name and tag, respectively, are the values of the `io.airbyte.name` and `io.airbyte.version` `LABEL`s in
 the Dockerfile.
 
 #### Run
+
 Then run any of the connector commands as follows:
+
 ```
-docker run --rm airbyte/source-shipstation-abstract:dev spec
-docker run --rm -v $(pwd)/secrets:/secrets airbyte/source-shipstation-abstract:dev check --config /secrets/config.json
-docker run --rm -v $(pwd)/secrets:/secrets airbyte/source-shipstation-abstract:dev discover --config /secrets/config.json
-docker run --rm -v $(pwd)/secrets:/secrets -v $(pwd)/integration_tests:/integration_tests airbyte/source-shipstation-abstract:dev read --config /secrets/config.json --catalog /integration_tests/configured_catalog.json
+docker run --rm airbyte/source-shipstation:dev spec
+docker run --rm -v $(pwd)/secrets:/secrets airbyte/source-shipstation:dev check --config /secrets/config.json
+docker run --rm -v $(pwd)/secrets:/secrets airbyte/source-shipstation:dev discover --config /secrets/config.json
+docker run --rm -v $(pwd)/secrets:/secrets -v $(pwd)/integration_tests:/integration_tests airbyte/source-shipstation:dev read --config /secrets/config.json --catalog /integration_tests/configured_catalog.json
 ```
+
 ## Testing
+
 Make sure to familiarize yourself with [pytest test discovery](https://docs.pytest.org/en/latest/goodpractices.html#test-discovery) to know how your test files and methods should be named.
 First install test dependencies into your virtual environment:
+
 ```
 pip install .[tests]
 ```
+
 ### Unit Tests
+
 To run unit tests locally, from the connector directory run:
+
 ```
 python -m pytest unit_tests
 ```
 
 ### Integration Tests
+
 There are two types of integration tests: Acceptance Tests (Airbyte's test suite for all source connectors) and custom integration tests (which are specific to this connector).
+
 #### Custom Integration tests
+
 Place custom tests inside `integration_tests/` folder, then, from the connector root, run
+
 ```
 python -m pytest integration_tests
 ```
+
 #### Acceptance Tests
+
 Customize `acceptance-test-config.yml` file to configure tests. See [Source Acceptance Tests](https://docs.airbyte.io/connector-development/testing-connectors/source-acceptance-tests-reference) for more information.
 If your connector requires to create or destroy resources for use during acceptance tests create fixtures for it and place them inside integration_tests/acceptance.py.
 To run your integration tests with acceptance tests, from the connector root, run
+
 ```
 python -m pytest integration_tests -p integration_tests.acceptance
 ```
+
 To run your integration tests with docker
 
 ### Using gradle to run tests
+
 All commands should be run from airbyte project root.
 To run unit tests:
+
 ```
-./gradlew :airbyte-integrations:connectors:source-shipstation-abstract:unitTest
+./gradlew :airbyte-integrations:connectors:source-shipstation:unitTest
 ```
+
 To run acceptance and custom integration tests:
+
 ```
-./gradlew :airbyte-integrations:connectors:source-shipstation-abstract:integrationTest
+./gradlew :airbyte-integrations:connectors:source-shipstation:integrationTest
 ```
 
 ## Dependency Management
+
 All of your dependencies should go in `setup.py`, NOT `requirements.txt`. The requirements file is only used to connect internal Airbyte dependencies in the monorepo for local development.
 We split dependencies between two groups, dependencies that are:
-* required for your connector to work need to go to `MAIN_REQUIREMENTS` list.
-* required for the testing need to go to `TEST_REQUIREMENTS` list
+
+- required for your connector to work need to go to `MAIN_REQUIREMENTS` list.
+- required for the testing need to go to `TEST_REQUIREMENTS` list
 
 ### Publishing a new version of the connector
+
 You've checked out the repo, implemented a million dollar feature, and you're ready to share your changes with the world. Now what?
+
 1. Make sure your changes are passing unit and integration tests.
 1. Bump the connector version in `Dockerfile` -- just increment the value of the `LABEL io.airbyte.version` appropriately (we use [SemVer](https://semver.org/)).
 1. Create a Pull Request.
