+++ airbyte-integrations/connectors/source-shipstation/acceptance-test-config.yml
@@ -1,9 +1,9 @@
 # See [Source Acceptance Tests](https://docs.airbyte.io/connector-development/testing-connectors/source-acceptance-tests-reference)
 # for more information about how to configure these tests
-connector_image: airbyte/source-shipstation-abstract:dev
+connector_image: airbyte/source-shipstation:dev
 tests:
   spec:
-    - spec_path: "source_shipstation_abstract/spec.yaml"
+    - spec_path: "source_shipstation/spec.yaml"
   connection:
     - config_path: "secrets/config.json"
       status: "succeed"
