+++ airbyte-integrations/connectors/source-shipstation/acceptance-test-docker.sh
