+++ airbyte-integrations/connectors/source-shipstation/integration_tests/abnormal_state.json
