+++ airbyte-integrations/connectors/source-shipstation/integration_tests/catalog.json
