+++ airbyte-integrations/connectors/source-shipstation/integration_tests/configured_catalog.json
