+++ airbyte-integrations/connectors/source-shipstation/integration_tests/invalid_config.json
