+++ airbyte-integrations/connectors/source-shipstation/integration_tests/sample_config.json
