+++ airbyte-integrations/connectors/source-shipstation/source_shipstation/schemas/customers.json
