+++ airbyte-integrations/connectors/source-shipstation/source_shipstation/schemas/get_store.json
