+++ airbyte-integrations/connectors/source-shipstation/source_shipstation/schemas/market_places.json
