+++ airbyte-integrations/connectors/source-shipstation/source_shipstation/schemas/stores.json
