+++ airbyte-integrations/connectors/source-shipstation/source_shipstation/schemas/warehouses.json
