+++ airbyte-integrations/connectors/source-shipstation/source_shipstation/spec.json
